refactor(work): extract shared readStateStep into hook-common (GH-313)
GH-313 added context-monitor.js with a readStateStep(tasksBase, ticket)
that was a near-identical copy of the same reader in capture-usage.js
(GH-311), tripping the static quality gate duplicate-blocks rule.

De-duplicate by hoisting the reader into the shared hook-common module
both hooks already import. The helper resolves the ticket id via
lib/config safeTicketId (fail-open to the raw id), reads .work-state.json,
and maps the 1-indexed currentStep to its ALL_STEPS name, returning
'unknown' on any missing or corrupt state. Behaviour is preserved exactly
in both callers; the now-unused fs import is dropped from capture-usage.js
while context-monitor.js keeps fs and safeTicket for its ledger paths.

Removes the non-allowlisted clone without growing .quality-exceptions.
